user class implemented

+++ src/main/java/io/moody/app/IDataAccessObject.java
@@ -3,7 +3,7 @@ package io.moody.app;
 public interface IDataAccessObject<T> {
     //crud - create, read, update, delete data from database;
     public boolean createRecord(T data);
-    public T readRecord(int id);
+    public T readRecord(long id);
     public boolean updateRecord(T data);
-    public boolean deleteRecord(int id);
+    public boolean deleteRecord(long id);
 }
